+++ src/cpp/core/include/core/json/Json.hpp
@@ -79,6 +79,10 @@ json::Array toJsonArray(const std::vector<T>& val)
    return results;
 }
 
+bool fillVectorString(const Array& array, std::vector<std::string>* pVector);
+bool fillVectorInt(const Array& array, std::vector<int>* pVector);
+bool fillMap(const Object& array, std::map< std::string, std::vector<std::string> >* pMap);
+
 bool parse(const std::string& input, Value* pValue);
 
 void write(const Value& value, std::ostream& os);