+++ src/cpp/r/include/r/RExec.hpp
@@ -300,7 +300,7 @@ private:
 
 
 class InterruptException {};
-   
+
 } // namespace exec   
 } // namespace r