+++ NEWS.md
@@ -13,6 +13,8 @@
     - Specialized autocompletions provided for `library`, `data`, `vignette`, `?`
     - For Shiny applications, autocompletions for `ui.R` + `server.R` pairs
     - Completions for dimension names in `[`, `[[` calls
+    - Completions from packages in `library`, `require` calls automatically
+      inferred and supplied even when not loaded
 * Improvements to C/C++ editing mode:
     - Code completion
     - F2 code navigation (go to definition)